+++ frontend/src/constants/chatbotConstants.js
@@ -1,3 +1,6 @@
 export const CHATBOT_CREATE_REQUEST = 'CHATBOT_CREATE_REQUEST';
 export const CHATBOT_CREATE_SUCCESS = 'CHATBOT_CREATE_SUCCESS';
 export const CHATBOT_CREATE_FAIL = 'CHATBOT_CREATE_FAIL';
+export const CHATBOT_UPDATE_REQUEST = 'CHATBOT_UPDATE_REQUEST';
+export const CHATBOT_UPDATE_SUCCESS = 'CHATBOT_UPDATE_SUCCESS';
+export const CHATBOT_UPDATE_FAIL = 'CHATBOT_UPDATE_FAIL';
